LayoutTests/imported/w3c/web-platform-tests/editing/other/delete-without-unwrapping-first-line-of-child-block_method=deleteCommand&lineBreak=br-expected.txt

Xdef
ghi

FAIL document.execCommand("delete") at abc<br>{<br><div id="child">]def<br>ghi</div> assert_in_array: document.execCommand("delete") at abc<br>{<br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc<br>def<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL document.execCommand("insertText", false, "X") at abc<br>{<br><div id="child">]def<br>ghi</div> assert_in_array: document.execCommand("insertText", false, "X") at abc<br>{<br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> value "abc<br>Xdef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">Xdef<br>ghi</div>", "abc<div id=\"child\">Xdef<br>ghi</div>"]
PASS document.execCommand("delete") at <br>[abc<br><div id="child">]def<br>ghi</div>
PASS document.execCommand("insertText", false, "X") at <br>[abc<br><div id="child">]def<br>ghi</div>
FAIL document.execCommand("delete") at <br>{<br><div id="child">]def<br>ghi</div> assert_equals: document.execCommand("delete") at <br>{<br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<br><div id=\"child\">def<br>ghi</div>" but got "<br>def<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("insertText", false, "X") at <br>{<br><div id="child">]def<br>ghi</div> assert_equals: document.execCommand("insertText", false, "X") at <br>{<br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<br><div id=\"child\">Xdef<br>ghi</div>" but got "<br>Xdef<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("delete") at {<br><br><div id="child">]def<br>ghi</div> assert_equals: document.execCommand("delete") at {<br><br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "def<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("insertText", false, "X") at {<br><br><div id="child">]def<br>ghi</div> assert_equals: document.execCommand("insertText", false, "X") at {<br><br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xdef<br>ghi</div>" but got "Xdef<br><div id=\"child\">ghi</div>"
PASS document.execCommand("delete") at [abc<br><br><div id="child">]def<br>ghi</div>
PASS document.execCommand("insertText", false, "X") at [abc<br><br><div id="child">]def<br>ghi</div>
FAIL document.execCommand("delete") at abc<br>{<br><div id="child">d]ef<br>ghi</div> assert_in_array: document.execCommand("delete") at abc<br>{<br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected text in the <div> value "abc<br>ef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">ef<br>ghi</div>", "abc<div id=\"child\">ef<br>ghi</div>"]
FAIL document.execCommand("insertText", false, "X") at abc<br>{<br><div id="child">d]ef<br>ghi</div> assert_in_array: document.execCommand("insertText", false, "X") at abc<br>{<br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> value "abc<br>Xef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">Xef<br>ghi</div>", "abc<div id=\"child\">Xef<br>ghi</div>"]
PASS document.execCommand("delete") at <br>[abc<br><div id="child">d]ef<br>ghi</div>
PASS document.execCommand("insertText", false, "X") at <br>[abc<br><div id="child">d]ef<br>ghi</div>
FAIL document.execCommand("delete") at <br>{<br><div id="child">d]ef<br>ghi</div> assert_equals: document.execCommand("delete") at <br>{<br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<br><div id=\"child\">ef<br>ghi</div>" but got "<br>ef<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("insertText", false, "X") at <br>{<br><div id="child">d]ef<br>ghi</div> assert_equals: document.execCommand("insertText", false, "X") at <br>{<br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<br><div id=\"child\">Xef<br>ghi</div>" but got "<br>Xef<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("delete") at {<br><br><div id="child">d]ef<br>ghi</div> assert_equals: document.execCommand("delete") at {<br><br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<div id=\"child\">ef<br>ghi</div>" but got "ef<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("insertText", false, "X") at {<br><br><div id="child">d]ef<br>ghi</div> assert_equals: document.execCommand("insertText", false, "X") at {<br><br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xef<br>ghi</div>" but got "Xef<br><div id=\"child\">ghi</div>"
PASS document.execCommand("delete") at [abc<br><br><div id="child">d]ef<br>ghi</div>
PASS document.execCommand("insertText", false, "X") at [abc<br><br><div id="child">d]ef<br>ghi</div>
FAIL document.execCommand("delete") at abc<br><br><div id="child">[]def<br>ghi</div> assert_in_array: document.execCommand("delete") at abc<br><br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc<br>def<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL document.execCommand("delete") at <br><br><div id="child">[]def<br>ghi</div> assert_equals: document.execCommand("delete") at <br><br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<br><div id=\"child\">def<br>ghi</div>" but got "<br>def<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("delete") at <br><div id="child">[]def<br>ghi</div> assert_equals: document.execCommand("delete") at <br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "def<br><div id=\"child\">ghi</div>"
FAIL document.execCommand("delete") at <b>abc<br><br></b></b><div id="child">[]def<br>ghi</div> assert_in_array: document.execCommand("delete") at <b>abc<br><br></b></b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> (<b> should stay) value "<b>abc<br></b>def<b><br></b><div id=\"child\">ghi</div>" not in array ["<b>abc<br></b><div id=\"child\">def<br>ghi</div>", "<b>abc</b><div id=\"child\">def<br>ghi</div>"]
PASS document.execCommand("delete") at <b><br></b><div id="child">[]def<br>ghi</div>
